Fix sqlx.Named to support IN queries and RETURNING
IN queries currently only work for scalar values and not tuples. This
allows for tuples with nested parens to be correctly parsed and
rewritten.

Previously only parens at the end of a query were considered for
rewriting. This allows for clauses like "RETURNING *" to be tacked onto
the end of a query, and instead only the last paren pair is considered
for rewriting.